requirements.txt
github3.py
mysql-connector-python

